fix(spreadsheet): rename sparkline helper to avoid case-insensitive filename clash (MED-295)
sparklineCell.ts (helpers) and SparklineCell.tsx (component) differ only by case, so on a case-insensitive filesystem (macOS) webpack resolved the helper import to the component module -> isSparklineRawInput undefined at runtime. Rename the helper module to sparklineData.ts (+ its test) so the names are distinct.
