+++ frontend/src/components/spreadsheets/sparklineData.ts
